Extract JSON-RPC response construction into JsonRpcResponseFactory
JsonRpcRequestProcessor mixed request parse/dispatch with success,
error, dispatch-accepted, heartbeat, and protocol-mismatch response
construction, plus 10 co-located response DTOs. Move all response
construction into a new stateless JsonRpcResponseFactory and split the
DTOs into class-per-file, so the processor is left owning only parse,
dispatch, notification, and cancellation handling.

CreateSuccessResponse/CreateErrorResponse/CreateCliProtocolMismatchResponse/
AppendTimingIfRequested widen from private to internal since the
processor now calls them across the class boundary; the two already-
internal CreateDispatchAcceptedResponse/CreateHeartbeatResponse and the
DTO visibility are unchanged. Refs: Refactor round 3 ToDo R3-5
