+++ Packages/src/Editor/Infrastructure/Api/CliUpdateRequiredErrorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3cc2486f961ea48df8310a6e0a6508a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
