+++ Packages/src/Editor/Infrastructure/Api/InternalErrorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf9c3c0f066f8433fb6c66f6f7666ae4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
