+++ Packages/src/Editor/Infrastructure/Api/JsonRpcError.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6642755ac33b843ad8b76cf7d190b473
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
