+++ Packages/src/Editor/Infrastructure/Api/JsonRpcErrorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e947996c8660428fbf126c017e3a21c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
