+++ Packages/src/Editor/Infrastructure/Api/JsonRpcErrorResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9fa02e1af45b74779be19677b0effd8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
