+++ Packages/src/Editor/Infrastructure/Api/JsonRpcErrorTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c51fd9b4f0b1f4afd8044c7aab0f1422
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
