+++ Packages/src/Editor/Infrastructure/Api/JsonRpcResponsePhases.cs
@@ -1,0 +1,8 @@
+namespace io.github.hatayama.UnityCliLoop.Infrastructure
+{
+    public static class JsonRpcResponsePhases
+    {
+        public const string Accepted = "accepted";
+        public const string Heartbeat = "heartbeat";
+    }
+}
