+++ Packages/src/Editor/Infrastructure/Api/JsonRpcResponsePhases.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0526adfb147024f46af583be80987102
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
