+++ Packages/src/Editor/Infrastructure/Api/JsonRpcSuccessResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9be8bf4a8ee444efa1adc14ccdf0794
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
