+++ Packages/src/Editor/Infrastructure/Api/SecurityBlockedErrorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 431d943a1f3f746279383089fbb7d729
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
