+++ Packages/src/Editor/Infrastructure/Api/ServerBusyErrorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd20357254e574ceabfbf2aa84f2a975
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
